+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed:
+- App icon on Android
+
+## [0.8.320] - 2023-04-15
+### Changed:
 - App icon on Windows
 - Upgraded dependencies
 - Improved color picker